Moved registrate method to abc level
