+++ common/python/src/common/messaging/clients.py
@@ -1,25 +1,13 @@
-import asyncio
 from aio_pika.abc import AbstractQueue, AbstractExchange
 import aio_pika
 
 from common.contracts.clients import MessageInClient, MessageOutClient
 
 
-def _is_async_callable(obj) -> bool:
-    return asyncio.iscoroutinefunction(obj) or (
-        callable(obj) and asyncio.iscoroutinefunction(obj.__call__)
-    )
-
-
 class RabbitMQInClient(MessageInClient):
     def __init__(self, queue: AbstractQueue):
         self._queue = queue
-        self._handler = None
-
-    def register(self, handler):
-        if not _is_async_callable(handler):
-            raise TypeError("Handler must be async.")
-        self._handler = handler
+        super().__init__()
 
     async def run(self):
         if self._handler is None:
